refactor: organize the Sources in different files for easier development
